pyproject.toml
[tool.poetry]
name = "sol_polykit"
version = "0.1.0"
description = ""
authors = ["Rishi Gurnani <rgurnani96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
rdkit-pypi = "^2021.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
